Add more routes using React Router
