+++ frontend/src/components/NewToy.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const NewToy = () => {
+  return (
+    <div>
+      <h1>
+        This page is going to create a new toy
+      </h1>
+    </div>
+  );
+};
+
+export default NewToy;
