+++ frontend/src/components/NotFound.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const NotFound = () => {
+
+  return (
+    <p>Page not found</p>
+  );
+};
+
+export default NotFound;
